Move job priority to enum file
