+++ server/src/jobs-new/index.ts
@@ -1,4 +1,5 @@
 import jobManager from './job-manager';
 
 export { JobType } from './job-type.enum';
+export { JobPriority } from './job-priority.enum';
 export { jobManager };
