+++ server/src/jobs-new/job-priority.enum.ts
@@ -1,0 +1,5 @@
+export enum JobPriority {
+  LOW = 3,
+  MEDIUM = 2,
+  HIGH = 1
+}
